JavaScript more modular, working on profile
* I seperated the javascript code into "shared" (for all pages), main (for posts page) and profile for the profile page
- Making it easier to maintain and more organized
* Updated html with new js scripts attached

* Split up SQLAlchemy models and Pydatic models (after advise from chatGPT)
- SQLAlchemy = DB logic
- Pydantic = API validation

* Added progress notes (to make some notes for documantation later on)
