+++ documentation/progress-notes.md
@@ -1,0 +1,3 @@
+14.05.25 - Kaja
+- Seperated SQLAlchemy models and Pydantic models to follow best practices (previously in the same file)
+    * SQLAlchemy = DB logic, Pydantic = API validation
