+++ frontend/frontend-scripts/shared.js
@@ -1,10 +1,35 @@
-import { getCachedPosts, cachePosts } from "./local-caching.js";
-
 const API_BASE_URL =
   
   // "https://idg2001-o1-social-networking-service.onrender.com"; // RENDER:
   "http://127.0.0.1:8000"; // LOCAL:
 
+// #####################
+// ### Local caching ###
+// #####################
+
+const CACHE_KEY = "cached_posts";
+const TIMESTAMP_KEY = `${CACHE_KEY}_timestamp`;
+const CACHE_EXPIRY = 5 * 60 * 1000; // 5 minutes in milliseconds
+
+export function getCachedPosts() {
+  const cached = localStorage.getItem(CACHE_KEY);
+  const timestamp = localStorage.getItem(TIMESTAMP_KEY);
+
+  if (!cached || !timestamp) return null;
+
+  const age = Date.now() - parseInt(timestamp);
+  if (age > CACHE_EXPIRY) {
+    // Cache too old.
+    return null;
+  }
+
+  return JSON.parse(cached);
+}
+
+export function cachePosts(posts) {
+  localStorage.setItem(CACHE_KEY, JSON.stringify(posts));
+  localStorage.setItem(TIMESTAMP_KEY, Date.now().toString());
+}
 
 // #######################
 // ### Event listeners ###
@@ -48,18 +73,6 @@ document.addEventListener("DOMContentLoaded", () => {
   }
 });
 
-// Welcome-message
-document.addEventListener("DOMContentLoaded", () => {
-  const welcomeMessage = document.querySelector("#welcome-message");
-
-  const userName = localStorage.getItem("user_name");
-  if (userName && welcomeMessage) {
-    welcomeMessage.textContent = `Welcome, ${userName}.`;
-  } else {
-    console.log("No 'user_name' found in storage.");
-  }
-});
-
 // Searchbar form (accessible and keyboard-friendly)
 document.addEventListener("DOMContentLoaded", () => {
   const searchForm = document.querySelector("#search-form");
@@ -71,13 +84,11 @@ document.addEventListener("DOMContentLoaded", () => {
   }
 });
 
-/////////////////////////////
 // Navbar User Toggle //
-///////////////////////////
 window.addEventListener("DOMContentLoaded", () => {
   const authLink = document.getElementById("auth-link"); 
 
-  fetch(`${API_BASE_URL}/auth/profile`, {
+  fetch(`${API_BASE_URL}/auth/me`, {
     credentials: "include", // sends the session_id cookie
   })
     .then((res) => {
@@ -101,13 +112,6 @@ window.addEventListener("DOMContentLoaded", () => {
     });
 });
 
-// Post-button
-document.addEventListener("DOMContentLoaded", () => {
-  const postBtn = document.querySelector("#post-button");
-  if (postBtn) {
-    postBtn.addEventListener("click", postPost);
-  }
-});
 
 // ########################
 // ### Global variables ###
@@ -283,132 +287,3 @@ function logout() {
       alert("Logout failed.");
     });
 }
-
-// Post a Cheep (Cheep a post? Cheep something?)
-function postPost() {
-  const content = document.getElementById("post-content").value;
-  const user_id = localStorage.getItem("user_id");
-
-  if (!user_id) {
-    alert("Please log in first!");
-    return;
-  }
-
-  fetch(`${API_BASE_URL}/posts/`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ user_id, content }),
-  })
-    .then((res) => res.json())
-    .then((data) => {
-      const msg = document.querySelector("#post-message");
-      const success = data.id && data.content;
-
-      msg.textContent = success
-        ? "Post created successfully!"
-        : data.detail || "Post failed.";
-      msg.className = success ? "success" : "error";
-
-      if (success) {
-        document.querySelector("#post-content").value = "";
-        loadPosts(); // Refresh the post list.
-      }
-    })
-    .catch((err) => console.error("Post error:", err));
-}
-
-// #################################
-// ### Loading & rendering posts ###
-// #################################
-// Load post feed (also integrate caching-functions)
-async function loadPosts() {
-  const postList = document.getElementById("post-list");
-
-  // Try to show cached posts (if valid).
-  const cached = getCachedPosts();
-  if (cached) {
-    console.log("Loaded posts from local cache.");
-    renderPosts(cached, postList);
-  }
-
-  // Always try to fetch fresh posts.
-  try {
-    const res = await fetch(`${API_BASE_URL}/posts/`);
-    if (!res.ok) throw new Error("Failed to fetch posts");
-
-    const posts = await res.json();
-
-    cachePosts(posts); // Update cache
-    console.log("Fetched fresh posts and updated cache.");
-    renderPosts(posts, postList); // Replace with fresh posts
-  } catch (err) {
-    console.error("Load posts error:", err);
-    if (!cached) {
-      postList.innerHTML =
-        "<p>Failed to load posts and no cached data available.</p>";
-    }
-  }
-}
-
-// For rendering posts
-function renderPosts(posts, container) {
-  container.innerHTML = "";
-
-  posts.forEach((post) => {
-    const postDiv = document.createElement("div");
-    postDiv.className = "post";
-
-    // Create like button
-    const likeBtn = document.createElement("button");
-    likeBtn.classList.add("like-btn");
-    likeBtn.textContent = post.is_liked_by_user ? "❤️" : "🤍";
-    likeBtn.style.marginLeft = "10px";
-
-    // Create like count
-    const likeCount = document.createElement("span");
-    likeCount.textContent = ` ${post.likes}`;
-    likeBtn.appendChild(likeCount);
-
-    // Like/unlike logic
-    likeBtn.addEventListener("click", async () => {
-      const token = localStorage.getItem("token");
-      const method = post.is_liked_by_user ? "DELETE" : "POST";
-      const url = `${API_BASE_URL}/posts/${post.id}/like`;
-
-      try {
-        res = await fetch(url, {
-          method: method,
-          credentials: "include", // Session-based cookies.
-        });
-
-        if (res.ok) {
-          post.is_liked_by_user = !post.is_liked_by_user;
-          post.likes += post.is_liked_by_user ? 1 : -1;
-
-          // Update like button display
-          likeBtn.innerHTML = post.is_liked_by_user ? "❤️" : "🤍";
-          likeCount.textContent = ` ${post.likes}`;
-          likeBtn.appendChild(likeCount);
-        } else {
-          alert("Failed to update like.");
-        }
-      } catch (err) {
-        console.error("Like button error:", err);
-        alert("Error with like button.");
-      }
-    });
-
-    postDiv.innerHTML = `
-      <strong>@${post.username || "anon"}</strong>: ${post.content}
-      <br><small>${new Date(post.timestamp).toLocaleString()}</small>
-    `;
-
-    // Append like button to post div
-    postDiv.appendChild(likeBtn);
-
-    container.appendChild(postDiv);
-  });
-}
-
-// Load posts on startup
-window.onload = loadPosts;
