+++ frontend/index.html
@@ -6,7 +6,8 @@
     <title>Cheeper</title>
     <link rel="stylesheet" href="style.css" />
     <link rel="icon" href="assets/cheeper-logo-blue-gradient.svg" />
-    <script src="frontend-scripts/script.js" type="module" defer></script>
+    <script src="frontend-scripts/shared.js" type="module" defer></script>
+    <script src="frontend-scripts/main-page.js" type="module" defer></script>
   </head>
   <body>
     <header>
