+++ frontend/profile.html
@@ -1,58 +1,54 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Profile</title>
-    <script
-      src="/frontend/frontend-scripts/profile.js"
-      type="module"
-      defer
-    ></script>
-    <script src="/frontend/frontend-scripts/script.js" defer></script>
-    <link rel="stylesheet" href="/frontend/style.css" />
-  </head>
-  <body>
-    <header>
-      <nav>
-        <div class="logo">
-          <img
-            class="logo__image"
-            src="assets/cheeper-logo-white.svg"
-            alt="The Cheeper logo visualizing a bird"
-          />
-          <h1 class="logo__text">Cheeper</h1>
-        </div>
-        <form id="search-form" role="search">
-          <label for="search-input" class="sr-only">Search posts</label>
-          <input type="text" id="search-input" placeholder="Search posts..." />
-          <button type="submit" id="search-button" aria-label="Search posts">
-            🔍
-          </button>
-        </form>
-        <button class="menu-toggle" id="menu-toggle">&#9776;</button>
-        <ul class="nav-links" id="nav-links">
-          <li><button id="dark-toggle">🌙</button></li>
-          <li><a id="auth-link" href="login_signup.html">Login</a></li>
-          <li><a href="index.html">Posts</a></li>
-        </ul>
-      </nav>
-    </header>
 
-    <div class="user-profile">
-      <div class="user-profile__info">
-        <img class="user-profile__info__user-image" />
-        <h2 class="user-profile__info__username"></h2>
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Profile</title>
+  <script src="frontend-scripts/shared.js" type="module" defer></script>
+  <script src="frontend-scripts/profile.js" type="module" defer></script>
+  <link rel="stylesheet" href="/frontend/style.css" />
+</head>
+
+<body>
+  <header>
+    <nav>
+      <div class="logo">
+        <img class="logo__image" src="assets/cheeper-logo-white.svg" alt="The Cheeper logo visualizing a bird" />
+        <h1 class="logo__text">Cheeper</h1>
       </div>
-      <div class="user-profile__my-posts">
-        <h2 class="user-profile__my-posts__header">My Posts</h2>
-        <div class="user-profile__my-posts__list"></div>
-      </div>
+      <form id="search-form" role="search">
+        <label for="search-input" class="sr-only">Search posts</label>
+        <input type="text" id="search-input" placeholder="Search posts..." />
+        <button type="submit" id="search-button" aria-label="Search posts">
+          🔍
+        </button>
+      </form>
+      <button class="menu-toggle" id="menu-toggle">&#9776;</button>
+      <ul class="nav-links" id="nav-links">
+        <li><button id="dark-toggle">🌙</button></li>
+        <li><a id="auth-link" href="login_signup.html">Login</a></li>
+        <li><a href="index.html">Posts</a></li>
+      </ul>
+    </nav>
+  </header>
+
+  <div class="user-profile">
+    <div class="user-profile__info">
+      <img class="user-profile__info__user-image" />
+      <h2 class="user-profile__info__username"></h2>
+      <h3 class="user-profile__info__email"></h3>
     </div>
-    <footer>
-      <button id="auth-button">Log Out</button>
-      <p id="auth-message" class="auth-message"></p>
-      <p>&copy; 2025 Cheeper. All rights reserved.</p>
-    </footer>
-  </body>
+    <div class="user-profile__my-posts">
+      <h2 class="user-profile__my-posts__header">My Posts</h2>
+      <div class="user-profile__my-posts__list"></div>
+    </div>
+  </div>
+  <footer>
+    <button id="auth-button">Log Out</button>
+    <p id="auth-message" class="auth-message"></p>
+    <p>&copy; 2025 Cheeper. All rights reserved.</p>
+  </footer>
+</body>
+
 </html>
